mmstudio: Various improvements to mmstudio
mmstudio: fix crash when closing ImageJ with closeOnExit = true

Some stuff was released but the Quit handler tried to release it again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

ZWheelListener: do not expect the zStageLabel to be null

MMCore cannot return a zStageLabel that is null, only an empty one.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

ReportingUtils: add a convenient logException() method

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

MMStudio: output more details on mmplugins to CoreLog

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

MMStudio: do not discover mmplugins/ wherever we might current be

Always discover mmplugins/ in the ImageJ directory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Never use the system class loader

We really have to keep in mind that it's not our business to expect a
certain way and only that way how to call Micro-Manager!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Report MMStartup errors much better

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Revert "Not sure if there are side-effects, but ignoring the focusLost event makes the property editor a lot nicer to work with"

There is a side effect. When typing a new value for a property and cliking
"Set" without pressing the Enter key first, the new value is completely
ignored... :-(

This reverts commit 2244c0874260011448e646d2a988d132382318e3.

Load AcqEngine classes with the same class loader as MMStudio itself

This should avoid class loading issues in case somebody loads the
studio using custom class loaders.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>