Hacked apart the demo camera a little to change image based on position
It isn't very pretty, but it finds a device named DXYStage and pulls the
position information from it, changing the image it generates. The image no
longer 'scrolls' (it's hard to check position when the image is constantly in
motion). Finally, the Y coordinate doesn't seem to do anything. I didn't think
on it long, but I couldn't come up with a reason for it.

Anyway, the point is so that my test acquisitions will be clearly varying in
position.