Threading and error handling revisions to performAcquisition.
I slightly reworked the way peformAndShowAcq functions. It now always produces
a new thread for performAcquisition (regardless of from whence it was called).
This thread is returned from the function, and should be interruptable. I also
made error handling a little more consistent; using performAndShowAcq, the
exceptions are now passed up the stack to the thread object, which displays a
message box and aborts. This is to ease the use of performAcquisition (and, to
a lesser degree, performAndShowAcq) by external libraries.