Added '3D' mode to DemoCamera (enabled via property browser).
I'm not entirely confident in the mathematics, but when enabled, the DemoCamera
will now project the corners of a cube as dots. They surround the 'stage'
positions and rotate around the 'upwards' axis much like the sample in the
OpenSPIM project does. It's still somewhat unstable, however; I've been
encountering various types of CTDs while it's enabled (still tracking them), so
play with it at your own risk. :) I've restored the original behavior of
DemoCamera, in favor of this approach.