Added interface (GUI) for a simpler n-dimensional acquisition sequence.
A hybrid of the SPIM controls (4D) and the tabular entry (nD). Users can select
any number of ranged devices (at the moment, they must be stages, though other
possibilities are easy enough to implement), then specify a minimum and maximum
value as well as a step size for each.