Fixed implementation of performAcquisition's timestep argument.
It used to be a delay (because that was the simpler thing to implement). Now,
the method tries to make each sequence start on time (such that it won't wait
at all if the sequence should've restarted already). This allows, conceptually,
something like "take a picture of this sample from every angle at 30-second
intervals".