Updated generateMeta to truncate decimals, and added some units.
When the stage is at 4000.00000000001, you get the point after the first four
zeroes. Floats and doubles in the generated metadata are now trimmed to four
decimal places. Time now explicitly has units of milliseconds (but it's a
double also, so its precision should still be sufficient), and X/Y stages now
have units of micrometers (which is what Micro-Manager calls for in the
device adapter implementation). Stages don't have any explicit units, since our
rotation axis theta is implemented as a stage and I can't really guess at the
units.