Fixed some image data conversion exceptions due to incorrect array types.
Micro-Manager apparently returns 32+ bit image data as an array of bytes (in
contrast to 16-bit, which is returned as a short[], and 8-bit, for which an
array of bytes is precisely what we want). My initial bToI function was not
only insufficient for floats/doubles, it was incorrect in its byte ordering.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>