Modified performAcquisition to indicate time dimension and open as hyperstack.
performAcquisition is taking multi-dimensional data; it follows that it should
let ImageJ *know* that the data is multi-dimensional. It now specifies the time
dimension as the number of frames, though multiple X/Y positions are indicated
along with Z positions as slices. (Separating these would require making the
method more aware of the actions it is performing; it really shouldn't have any
more idea than it does already.)

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>