Added min/max properties for X/Y stage and corrected orientation issues.
The Picard X/Y stage motors were reacting backwards from what we wanted,
and they also had no well-established and enforced minimum/maximum. I've
added these as properties in the hardware config.

A quick note on the logic of GetOrientation in OpenSPIM.cpp: Since our
motors were reacting backwards by default, I have the function default to
mirroring X/Y. The end result is that, with the mirror X/Y properties both
set to zero, the stage moves as we expect it to. Changing these to 1 flips
the directions they move in.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>