Implemented rudimentary 'rotation' in the live window.
By holding alt and dragging, you can now rotate the twister motor. It
isn't the full implementation, in which we hope to rotate around a
specific axis in 3D space (involves running the stage motors also), but
it's proof that we can hook into the window. :)

(Sort of) Signed-off-by: Luke Stuyvenberg (stuyvenberg@wisc.edu)