Added the ability to guess a third point.
As a start of automatic calibration, I implemented a 'guess the third
position'. After selecting two points during calibration, the user can
click 'guess' to try and reposition the stage to show where the third
point should be, using an approximation of the axis as vertical. It
doesn't seem to guess right yet, and I'm not sure how to improve the
accuracy except possibly by wider angles. Requires further testing.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>