Further revisions to OME-TIFF handler.
I can't seem to win.

If I explicity specify the plane count, it gets overwritten by the value
1. See OMETiffWriter.java:289. I take this to mean each TiffData should
only have one plane specified in it (despite a PlaneCount parameter?).
Worth noting is that if I correct the plane count in this case, the images
are imported correctly.

If I attempt to follow the guideline given there -- i.e. one plane per
TiffData -- all TiffData elements past the first in each Image element have an
incorrect UUID (I suspect my decided-upon UUID is replaced anyway, but in
the first case, there's only one TiffData element, so it's replaced
_correctly_.). See OMETiffWriter:342.

If I remove the UUID value setting and rely on the writer
to handle it for me (as the fixup to this commit will be written to do),
the second time point for each angle (i.e. the sixth TiffData element,
given a dataset produced by OMETIFFHandler's main()) is missing a new UUID
for its new filename (contrary to the way I interpreted the source code;
not sure why yet!).

Essentially, bio-formats seems intent on being as helpful as possible,
whether I want it to be or not. ;)

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>