[0000058] Auto-detecting Pi motors/twisters
When the DAL is initialized by Micro-Manager, it pings all motor and
twister indices between 0 and 500 (inclusive), storing each hit (each
motor/twister that connected successfully) in a list. This list is then
consulted by the stages and twister -- the first 3 motors will be assigned
Y/X/Z, and the first twister will be the rotation. If there are less than
3 motors or no twisters, those missing serial numbers will be replaced
with -1, but may be changed by the user (so long as they can initialize
the associated motor).

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>