Batch_Convert_Movies: support MacOSX
This is a companion commit to fiji.git's 2e33a490(Make_Screencast.bsh:
also search for VLC at its default location on MacOSX) by Jan Eglinger.

Unfortunately, it seems as if current VLC (2.0.5) segfaults when called on
the command-line on MacOSX 10.5, as reported by Pavel Tomancak, although
it worked fine on MacOSX 10.6.8.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>