Prepare add-desktop-shortcut to run outside msysGit
We want to switch the development environment from an msysGit-based one to a Git for Windows one. (msysGit is the development environment for Git for Windows, featuring a GNU C/C++ compiler, and building Git from scratch, which we do not need here.)

The original idea was that we need the C compiler to build Fiji, which was true until the Fiji launcher migrated to ImageJ2. But we do not even need to build Fiji any longer; Jenkins does that for us.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>