Made long-waits less intrusive
Look, they're still necessary. If the motors don't move but Micro-Manager
thinks they have, some really _weird_ stuff is going to happen. But per
Johannes' suggestion, I've limited the maximum waiting time -- in this
case, to 300 ms. Well, technically 360, but it's pretty close, isn't it?

They still happen -- a LOT -- when the Z motor is scrolled while the pixel
size is calibrated. This is because Micro-Manager thinks the motor can
move fewer than 0.75 um (it can't). And if you scroll a lot, it'll queue a
bunch up and you'll overshoot. Still, better than it used to be.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>