Rebuilt acquisition model using device manager
Huge change. This touches basically every file involved in progacq, but I
think it's worth it.

So, what changed? It's probably easiest to start with the StepTableModel:
it now describes a table of AcqRows, not String[]s (ick!). This cleans up
the pre-acquisition setup (no converting rows; we just borrow the
list...).

Accordingly, AcqRow now uses the DeviceManager's device enum to map
devices to their 'positions'. This also eliminates some nasty
stringification (the positions are stored as doubles now...).

This led to a tangential change in the ProjDiffAntiDrift (just to adapt to
some changes in method names).

AcqParams also needed to change in order to carry a DeviceManager to
ProgrammaticAcquisitor (so it can look up these device labels).

Obviously, this means that ProgrammaticAcquisitor had to change. Most of
these are pretty tame changes, just sorta hacking the new way of referring
to devices into the code -- it's not very pretty (the DeviceManager gets
passed around a lot now...), but it lets me refer to specific devices of a
setup in a general way (i.e. 'send the Z-stage to some-coordinate', or
'get the position of the theta-stage').

Finally, of course, SPIMAcquisition needed some changing also. The
aforementioned pre-acquisition cleanup, mostly.

The drift correction now propagates back to the position list. This is a
consequence of the same AcqRow's modifying themselves to correct for
drift. I personally regard this as a plus -- if you need to reset the
acquisition, any initial manual correction and any subsequent drift
corrections will already be applied to the positions. If it really needs
to change, we only need to deep-copy the AcqRows.

So yes, a lot of changes with no visible difference, but I'm confident it
will have its advantages.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>