Added experimental Picard metastage device
Essentially, this implements the idea I had in my comment on issue 120 on
the bugtracker. The new 'device' creates two CSIABStage instances and
bundles them as an X/Y stage.

I'd like to stress that this is experimental, and shouldn't be relied upon
for anything yet. In fact, prior to any uploads of the PicardStage DAL, it
may be wise to comment out the 'AddAvailableDevice' call so it doesn't
show up.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>