SteppedSlider enhancements and fixes
trySetValue no longer goes through with it if the value is the same. This
needs to happen in some places, but trySetValue -- used by the timer motor
updates in SPIMAcquisition -- isn't one of them, and it was interfering
with the ability to enter and go to values.

I got rid of the horizontal struts in the value/range box at the bottom of
the slider; for some reason, these were increasing the box height
significantly. If they look too awful, I'll put in a 4x1 fixed box...

The initial value of a DoubleField is now rounded as well. A minor fix,
but it annoyed me.

I added functions for getting the current minimum and maximum of the
slider. Still no way to set the rounding. :(

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>