Async handler: Reworked & added monitor
Reworked async output handler

It now queues stack begin/end events, to ensure they don't happen out of
order with the slice writes, and to save us the problem of (very
synchronously) emptying the queue in between each view.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>

Added async monitor

In effort to debug the asynchronous writing (or at least gain some insight
into its present state), I added a monitor which graphs the size of the
event/slice queue over time and mentions when it is writing things (sort
of).

So far, I have learned that it writes things pretty quickly. The highest
queue count I saw so far was 5 (out of 228), and this was while I had
another program scouring the hard drive in the background. Prime95 doesn't
even seem to faze it. I may make an effort to slow down or otherwise delay
the writing -- it is still using time which might be best spent during
acquisition (or perhaps not? Perhaps the time it is borrowing all comes
from waits and sleeps on the acquisition thread. It warrants investigation.)

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>

Improved async monitor implementation

The monitoring process is now a runnable, instantiated by the wrapper (and
not the writer thread). The writing variable is also controlled in the
wrapper (so it doesn't flicker so much).

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>