Added async monitor
In effort to debug the asynchronous writing (or at least gain some insight
into its present state), I added a monitor which graphs the size of the
event/slice queue over time and mentions when it is writing things (sort
of).

So far, I have learned that it writes things pretty quickly. The highest
queue count I saw so far was 5 (out of 228), and this was while I had
another program scouring the hard drive in the background. Prime95 doesn't
even seem to faze it. I may make an effort to slow down or otherwise delay
the writing -- it is still using time which might be best spent during
acquisition (or perhaps not? Perhaps the time it is borrowing all comes
from waits and sleeps on the acquisition thread. It warrants investigation.)

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>