Stopped wild deletion of destination directories
Before, the plugin would delete any and all files it could under the
destination directory. In hindsight, this might've caused enormous grief
to someone unprepared, as the message gave no indication this would occur.

To remedy this, I added a warning to the message, and filtered the files
by filename -- the plugin will delete all .ome.tiff in the destination
directory, but nothing else. This may need revision if another output
handler is ever made, unlikely though it may be.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>