Long-wait improvements
I was foolishly making use of time(NULL) instead of clock() for
long-waits. As a result, the duration of waiting depended strongly on how
close to a second-barrier your computer clock was.

I've switched the waits to clock(), which should be more precise (time is
in seconds, clock is system-dependent but often in millis). I also made
their messages debug-only (and provided more information in them).

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>