Autocalibrator fitting & interface improvements
The circle fitting for the autocalibrator has been completely overhauled.
It now implements a particular case of the more general hypersphere fitter
(that is, one in *TWO* dimensions), and also fits the rotational axis
normal. Wildly eccentric axes will still throw it off, but that's no
surprise.

The biggest change besides the above is that the autocalibrator will wrap
around when it reaches +180, rather than going off into >180 territory
which confuses the heck out of the plugin.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>