PicardStage: use RegisterDevice; spaces -> tabs
The MM API will be changing to require RegisterDevice instead of
AddAvailableDeviceName, so the core can find out the type of a device
without instantiating it. I've replaced our calls and put in the correct
device types.

While I was at it (or, I guess, vice-versa), I changed the inconsistent
spacing-indentation with tabs. Curiously, while PicardStage.h had a
four-space indent (ick, but okay, I guess...), PicardStage.cpp had a
three-space indent (Three? Seriously? Four I can understand, two in a
pinch, but three!?). But it's okay! They're all tabs now.

Finally, I rearranged some bits of the solution file and added some
missing lines that may have stopped the project from building in 64-bit
mode. Most of this is in preparation for submission to Micro-Manager.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>